requirements.txt
flask==3.0.0
gunicorn==21.2.0
psycopg[binary]>=3.2.0
leash-sdk>=0.3.0
